requirements.txt
requests>=2.28.0
pillow>=9.0.0
pyyaml>=6.0
python-dotenv>=0.20.0
tqdm>=4.64.0
argparse>=1.4.0
loguru>=0.6.0
base64io>=1.0.3
pyjwt>=2.4.0
